+++ R/Learner.R
@@ -184,7 +184,7 @@ Learner = R6Class("Learner",
     #' @param ... (ignored).
     print = function(...) {
       catn(format(self), if (is.null(self$label) || is.na(self$label)) "" else paste0(": ", self$label))
-      catn(str_indent("* Model:", if (is.null(self$model)) "-" else class(self$model)[1L]))
+      catn(str_indent("* Model:", if (is.null(self$model)) "-" else if (marshalled_model(self$model)) "<marshalled>" else paste0(class(self$model)[1L])))
       catn(str_indent("* Parameters:", as_short_string(self$param_set$values, 1000L)))
       catn(str_indent("* Packages:", self$packages))
       catn(str_indent("* Predict Types: ", replace(self$predict_types, self$predict_types == self$predict_type, paste0("[", self$predict_type, "]"))))
@@ -277,6 +277,10 @@ Learner = R6Class("Learner",
 
       if (is.null(self$state$model) && is.null(self$state$fallback_state$model)) {
         stopf("Cannot predict, Learner '%s' has not been trained yet", self$id)
+      }
+
+      if (marshalled_model(self$model)) {
+        stopf("Cannot predict, Learner '%s' has not been unmarshalled yet", self$id)
       }
 
       if (isTRUE(self$parallel_predict) && nbrOfWorkers() > 1L) {
@@ -387,7 +391,6 @@ Learner = R6Class("Learner",
       }
       self$state$model
     },
-
 
     #' @field timings (named `numeric(2)`)\cr
     #' Elapsed time in seconds for the steps `"train"` and `"predict"`.
@@ -540,7 +543,6 @@ Learner = R6Class("Learner",
     }
   )
 )
-
 
 #' @export
 rd_info.Learner = function(obj, ...) {