+++ tests/testthat/test_HotstartStack.R
@@ -409,3 +409,11 @@ test_that("HotstartStack threshold works", {
   hot$add(learner_2)
   expect_data_table(hot$stack, nrows = 1)
 })
+
+test_that("error when adding marshalled learner", {
+  hot = HotstartStack$new()
+  learner = lrn("classif.lily")
+  learner$train(tsk("iris"))
+  learner$marshal()
+  expect_error(hot$add(learner), "unmarshalled")
+})